refactor: refactors the Spinner component to the new code style
